// apps/v2/src/app/services/sdk/types.gen.ts
export type MovieDto = {
  _id: string;
  title: string;
  /**
   * Original title
   */
  originalTitle: string;
  releaseDate: string;
  /**
   * Overview
   */
  overview: string;
  posterUrl: string;
  backdropUrl: string;
  adult: boolean;
};

export type MovieCommentDto = {
  _id: string;
  /**
   * Rating
   */
  rating: number;
  userId: string;
  userName: string;
  movieId: string;
  content: string;
};

export type CreateMovieCommentDto = {
  /**
   * Rating
   */
  rating: number;
  movieId: string;
  content: string;
};